Revert "Usa ints ao inves de BigIntegers - #443"
This reverts commit d8012374dbd4375f7d072d76955faf7f8cc15eab.

+++ src/main/java/br/gov/servicos/v3/schema/Ate.java
@@ -8,6 +8,7 @@
 
 package br.gov.servicos.v3.schema;
 
+import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
@@ -37,23 +38,31 @@ import javax.xml.bind.annotation.XmlType;
 public class Ate {
 
     @XmlAttribute(name = "max", required = true)
-    protected int max;
+    protected BigInteger max;
     @XmlAttribute(name = "unidade", required = true)
     protected TipoDeTempoTotalEstimado unidade;
 
     /**
      * Gets the value of the max property.
      * 
+     * @return
+     *     possible object is
+     *     {@link BigInteger }
+     *     
      */
-    public int getMax() {
+    public BigInteger getMax() {
         return max;
     }
 
     /**
      * Sets the value of the max property.
      * 
+     * @param value
+     *     allowed object is
+     *     {@link BigInteger }
+     *     
      */
-    public void setMax(int value) {
+    public void setMax(BigInteger value) {
         this.max = value;
     }
